move leverage calculation to components
