+++ components/trade/PerpButtonGroup.tsx
@@ -1,24 +1,52 @@
+import { PerpMarket } from '@blockworks-foundation/mango-v4'
 import ButtonGroup from '@components/forms/ButtonGroup'
 import mangoStore from '@store/mangoStore'
-import { useCallback, useState } from 'react'
+import useMangoAccount from 'hooks/useMangoAccount'
+import useSelectedMarket from 'hooks/useSelectedMarket'
+import { useCallback, useMemo, useState } from 'react'
 import { trimDecimals } from 'utils/numbers'
 
 const PerpButtonGroup = ({
-  max,
   minOrderDecimals,
   tickDecimals,
 }: {
-  max: number
   minOrderDecimals: number
   tickDecimals: number
 }) => {
+  const side = mangoStore((s) => s.tradeForm.side)
+  const { selectedMarket } = useSelectedMarket()
+  const { mangoAccount } = useMangoAccount()
   const [sizePercentage, setSizePercentage] = useState('')
+  const tradeFormPrice = mangoStore((s) => s.tradeForm.price)
+
+  const leverageMax = useMemo(() => {
+    const group = mangoStore.getState().group
+    if (!mangoAccount || !group || !selectedMarket) return 100
+    if (!(selectedMarket instanceof PerpMarket)) return 100
+
+    try {
+      if (side === 'buy') {
+        return mangoAccount.getMaxQuoteForPerpBidUi(
+          group,
+          selectedMarket.perpMarketIndex
+        )
+      } else {
+        return mangoAccount.getMaxBaseForPerpAskUi(
+          group,
+          selectedMarket.perpMarketIndex
+        )
+      }
+    } catch (e) {
+      console.error('Error calculating max leverage perp btn grp: ', e)
+      return 0
+    }
+  }, [side, selectedMarket, mangoAccount, tradeFormPrice])
 
   const handleSizePercentage = useCallback(
     (percentage: string) => {
       const set = mangoStore.getState().set
       setSizePercentage(percentage)
-      const size = max * (Number(percentage) / 100)
+      const size = leverageMax * (Number(percentage) / 100)
 
       set((s) => {
         if (s.tradeForm.side === 'buy') {
@@ -48,7 +76,7 @@ const PerpButtonGroup = ({
         }
       })
     },
-    [max, minOrderDecimals, tickDecimals]
+    [leverageMax, minOrderDecimals, tickDecimals]
   )
 
   return (
